@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
-import { mergeMap, switchMap } from 'rxjs/operators';
-import { setNewRestaurant } from '../../restaurant/store/restaurant.actions';
+import { debounceTime, mergeMap, switchMap } from 'rxjs/operators';
+import { validateNameRestaurant, setNewRestaurant, setNameError, setNoNameError } from '../../restaurant/store/restaurant.actions';
 import { RestaurantApiService } from '../services/restaurant-api.service';
 import { createRestaurant } from './restaurant.actions';
 
@@ -14,6 +14,24 @@ export class RestaurantEffects {
         return this.restaurantApiService.postRestaurant(action).pipe(
           switchMap((restaurant) => {
             return [setNewRestaurant({ restaurant })];
+          })
+        );
+      })
+    )
+  );
+
+  validateNameRestaurant$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(validateNameRestaurant),
+      debounceTime(500),
+      mergeMap((action: { name: string }) => {
+        return this.restaurantApiService.validateRestaurantName(action.name).pipe(
+          switchMap((isValid: boolean) => {
+            if (isValid) {
+              return [setNoNameError()];
+            } else {
+              return [setNameError({ error: true, duplicated: true })];
+            }
           })
         );
       })
